elearning 95% and add config.js for global env

+++ e-learning/resources/js/components/AnnouncementComponent.vue
@@ -29,7 +29,7 @@
 											<th>#</th>
 											<th>Kelas</th>
 											<th>Pesan</th>
-											<th>Muncul Sampai</th>
+											<th>Batas Pengumuman</th>
 											<th></th>
 										</tr>
 									</thead>
@@ -81,7 +81,7 @@
 									<textarea class="form-control" v-model="data.messages" cols="30" rows="10"></textarea>
 								</div>
 								<div class="form-group">
-									<label>Muncul Sampai</label>
+									<label>Batas Pengumuman</label>
 									<input type="datetime-local" class="form-control" v-model="data.expired_at" />
 								</div>
 							</div>
@@ -130,7 +130,7 @@ export default {
             'Content-Type':'application/json'
         }
         if(this.token === undefined || this.token === null || this.token === '' ){
-            window.location = process.env.MIX_ES_URL+'/login'
+            window.location = window.config.MIX_ES_URL+'/login'
         }
         await this.fetchUserId()
         await this.findEmployee()
@@ -140,7 +140,7 @@ export default {
 
 		// ANNOUNCEMENT
 		async fetchUserId(){
-            let response = await fetch(process.env.MIX_ES_URL+'/api/details',{
+            let response = await fetch(window.config.MIX_ES_URL+'/api/details',{
                 method:'post',
                 headers:this.headers
             });
@@ -160,6 +160,11 @@ export default {
 				val.classroom = obj
 	        })
 	        this.announcements = await data
+
+	        setTimeout(()=>{
+                vm.status = false
+                vm.deleteStatus = false
+            },2500)
 	        return data
         },
         findAnnouncement(id){
@@ -199,21 +204,40 @@ export default {
             })
         },
         deleteAnnouncement(announcement_id){
-            fetch('api/announcement/delete',{
-                method:'delete',
-                headers:this.headers,
-                body:JSON.stringify({id:announcement_id,teacher_id:this.other_id})
-            })
-            .then(res=>res.json())
-            .then(res=>{
-                this.deleteStatus = true
-                this.loadAnnouncements()
-            })
+        	var vm = this
+        	Swal.fire({
+			  title: 'Apakah anda yakin akan menghapus data ini?',
+			  text: "Tindakan ini tidak dapat dikembalikan!",
+			  type: 'warning',
+			  showCancelButton: true,
+			  confirmButtonColor: '#3085d6',
+			  cancelButtonColor: '#d33',
+			  cancelButtonText: 'Tidak',
+			  confirmButtonText: 'Ya, Hapus data ini!'
+			}).then((result) => {
+			  if (result.value) {
+			  	fetch('api/announcement/delete',{
+	                method:'delete',
+	                headers:vm.headers,
+	                body:JSON.stringify({id:announcement_id,teacher_id:vm.other_id})
+	            })
+	            .then(res=>res.json())
+	            .then(res=>{
+	                vm.deleteStatus = true
+	                vm.loadAnnouncements()
+	                Swal.fire(
+				      'Terhapus!',
+				      'Data berhasil dihapus.',
+				      'success'
+				    )
+	            })
+			  }
+			})
 		},
 
 		// /ANNOUNCEMENT
 		loadEmployees(){
-			fetch(process.env.MIX_IS_URL+'/api/employee',{
+			fetch(window.config.MIX_IS_URL+'/api/employee',{
                 headers:this.headers,
             })
             .then(res => res.json())
@@ -222,7 +246,7 @@ export default {
             })
 		},
 		async findEmployee(){
-			let response = await fetch(process.env.MIX_IS_URL+'/api/employee/'+this.other_id,{
+			let response = await fetch(window.config.MIX_IS_URL+'/api/employee/'+this.other_id,{
                 headers:this.headers,
             });
             let data = await response.json()
