+++ e-learning/resources/js/components/VirtualClassroomComponent.vue
@@ -194,7 +194,7 @@ export default {
             'Content-Type':'application/json'
         }
         if(this.token === undefined || this.token === null || this.token === '' ){
-            window.location = process.env.MIX_ES_URL+'/login'
+            window.location = window.config.MIX_ES_URL+'/login'
         }
         await this.fetchUserId()
         await this.findEmployee()
@@ -204,7 +204,7 @@ export default {
 
 		// ANNOUNCEMENT
 		async fetchUserId(){
-            let response = await fetch(process.env.MIX_ES_URL+'/api/details',{
+            let response = await fetch(window.config.MIX_ES_URL+'/api/details',{
                 method:'post',
                 headers:this.headers
             });
@@ -274,21 +274,40 @@ export default {
             })
         },
         deleteVirtualClassroom(virtual_classroom_id){
-            fetch('api/exam/delete',{
-                method:'delete',
-                headers:this.headers,
-                body:JSON.stringify({id:virtual_classroom_id,teacher_id:this.other_id})
-            })
-            .then(res=>res.json())
-            .then(res=>{
-                this.deleteStatus = true
-                this.loadVirtualClassrooms()
-            })
+        	var vm = this
+        	Swal.fire({
+			  title: 'Apakah anda yakin akan menghapus data ini?',
+			  text: "Tindakan ini tidak dapat dikembalikan!",
+			  type: 'warning',
+			  showCancelButton: true,
+			  confirmButtonColor: '#3085d6',
+			  cancelButtonColor: '#d33',
+			  cancelButtonText: 'Tidak',
+			  confirmButtonText: 'Ya, Hapus data ini!'
+			}).then((result) => {
+			  if (result.value) {
+			  	fetch('api/exam/delete',{
+	                method:'delete',
+	                headers:vm.headers,
+	                body:JSON.stringify({id:virtual_classroom_id,teacher_id:vm.other_id})
+	            })
+	            .then(res=>res.json())
+	            .then(res=>{
+	                vm.deleteStatus = true
+	                vm.loadVirtualClassrooms()
+	                Swal.fire(
+				      'Terhapus!',
+				      'Data berhasil dihapus.',
+				      'success'
+				    )
+	            })
+			  }
+			})
 		},
 
 		// /ANNOUNCEMENT
 		loadEmployees(){
-			fetch(process.env.MIX_IS_URL+'/api/employee',{
+			fetch(window.config.MIX_IS_URL+'/api/employee',{
                 headers:this.headers,
             })
             .then(res => res.json())
@@ -297,7 +316,7 @@ export default {
             })
 		},
 		async findEmployee(){
-			let response = await fetch(process.env.MIX_IS_URL+'/api/employee/'+this.other_id,{
+			let response = await fetch(window.config.MIX_IS_URL+'/api/employee/'+this.other_id,{
                 headers:this.headers,
             });
             let data = await response.json()
